+++ crates/tx-relay/src/evm/vanchor.rs
@@ -4,6 +4,7 @@ use crate::TransactionItemKey;
 use ethereum_types::{H512, U256};
 use futures::TryFutureExt;
 use std::{collections::HashMap, sync::Arc};
+use webb::evm::ethers::prelude::Middleware;
 use webb::evm::ethers::types::transaction::eip2718::TypedTransaction;
 use webb::evm::ethers::utils::{format_units, hex, parse_ether};
 use webb::evm::{
